 requirements.txt
progress.bar
requests